# docs/swagger.yaml
basePath: /
definitions:
  models.ApiError:
    properties:
      error:
        type: string
    type: object
...
  models.ApiMessage:
    properties:
      message:
        type: string
    type: object
...
  models.UserEmail:
    properties:
      email:
        example: nktkln@example.com
        type: string
    type: object
...
  models.UserName:
    properties:
      name:
        example: NKTKLN
        type: string
    type: object
  models.UserPassword:
    properties:
      password:
        example: StRon9Pa$$w0rd
        type: string
    type: object
  models.UserTokens:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIzMDE5NDgxODQiLCJleHAiOjE2NTU0MTc4MDF9.ZqPxmPK3qV3VrT4D0wbwN2tMGzAhaH5kQnqr8iePTZA
        type: string
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIzMDE5NDgxODQiLCJleHAiOjE2NTU0MjI1NTh9.S1ypUshOnPB66VJV6RL42cdgYbV8LaGk7zYgL5JlsYg
        type: string
    type: object
  models.UserUsername:
    properties:
      username:
        example: nktkln
        type: string
    type: object
info:
  contact:
    email: nktkln@nktkln.com
    name: API Support
    url: https://nktkln.com
  description: some description
  license:
    name: MIT
    url: http://github.com/
  title: ToDo API
  version: "1.0"
paths:
...
  /user/delete/account:
    delete:
      consumes:
      - application/json
      parameters:
      - description: User password
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApiMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      security:
      - token: []
      summary: Delete user account
      tags:
      - User settings
  /user/delete/icon:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApiMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      security:
      - token: []
      summary: Delete user icon
      tags:
      - User settings
  /user/settings/reset/email:
    post:
      consumes:
      - application/json
      parameters:
      - description: New user email
        in: body
        name: NewUserEmail
        required: true
        schema:
          $ref: '#/definitions/models.UserEmail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApiMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      security:
      - token: []
      summary: Reset user email
      tags:
      - User settings
  /user/settings/reset/password:
    post:
      consumes:
      - application/json
      parameters:
      - description: User data
        in: body
        name: UserEmail
        required: true
        schema:
          $ref: '#/definitions/models.UserEmail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserTokens'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Reset user password
      tags:
      - User settings
  /user/settings/update/email:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Veryfication key
        in: query
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApiMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Update user email
      tags:
      - User settings
  /user/settings/update/icon:
    put:
      consumes:
      - application/json
      parameters:
      - description: User icon
        in: formData
        name: icon
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApiMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      security:
      - token: []
      summary: Update user icon
      tags:
      - User settings
  /user/settings/update/name:
    patch:
      consumes:
      - application/json
      parameters:
      - description: User name
        in: body
        name: NewUserName
        required: true
        schema:
          $ref: '#/definitions/models.UserName'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApiMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      security:
      - token: []
      summary: Change user name
      tags:
      - User settings
  /user/settings/update/password:
    patch:
      consumes:
      - application/json
      parameters:
      - description: User password
        in: body
        name: NewUserPassword
        required: true
        schema:
          $ref: '#/definitions/models.UserPassword'
      - description: Veryfication key
        in: query
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserTokens'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Update user password
      tags:
      - User settings
  /user/settings/update/token:
    put:
      consumes:
      - application/json
      parameters:
      - description: Refresh token
        in: header
        name: refresh_token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserTokens'
      summary: Update user token
      tags:
      - User settings
  /user/settings/update/username:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Username
        in: body
        name: NewUsername
        required: true
        schema:
          $ref: '#/definitions/models.UserUsername'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApiMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      security:
      - token: []
      summary: Change user username
      tags:
      - User settings
...
securityDefinitions:
  token:
    in: header
    name: token
    type: apiKey
swagger: "2.0"